refactor: refactor terminal components
