+++ packages/ui/src/app/centurion-app.tsx
@@ -1,12 +1,13 @@
 import { ClientAPI } from "@rbxts/centurion";
-import { UserInputService } from "@rbxts/services";
+import { GuiService, UserInputService } from "@rbxts/services";
 import Vide, { derive } from "@rbxts/vide";
-import Terminal from "../components/terminal";
+import { Suggestions, Terminal } from "../components";
+import { Group } from "../components/ui/group";
 import { Layer } from "../components/ui/layer";
 import { useAPI } from "../hooks/use-api";
 import { useAtom } from "../hooks/use-atom";
 import { useEvent } from "../hooks/use-event";
-import { usePx } from "../hooks/use-px";
+import { px, usePx } from "../hooks/use-px";
 import {
 	interfaceOptions,
 	interfaceVisible,
@@ -19,7 +20,7 @@ const MOUSE_INPUT_TYPES = new Set<Enum.UserInputType>([
 	Enum.UserInputType.Touch,
 ]);
 
-export function TerminalApp(api: ClientAPI) {
+export function CenturionApp(api: ClientAPI) {
 	useAPI(api);
 	usePx();
 
@@ -42,7 +43,19 @@ export function TerminalApp(api: ClientAPI) {
 
 	return (
 		<Layer displayOrder={() => options().displayOrder ?? 0} visible={visible}>
-			<Terminal />
+			<Group
+				anchorPoint={() => options().anchorPoint ?? new Vector2(0, 0)}
+				size={() => options().size ?? new UDim2(0, px(1024), 1, 0)}
+				position={() =>
+					options().position ??
+					UDim2.fromOffset(px(16), px(8) + GuiService.GetGuiInset()[0].Y)
+				}
+			>
+				<Terminal />
+				<Suggestions />
+
+				<uilistlayout Padding={new UDim(0, px(8))} SortOrder={"LayoutOrder"} />
+			</Group>
 		</Layer>
 	);
 }
