+++ packages/ui/src/components/index.ts
@@ -1,1 +1,2 @@
+export * from "./suggestions";
 export * from "./terminal";
