+++ packages/ui/src/components/terminal/index.ts
@@ -1,0 +1,1 @@
+export * from "./terminal";
